proxy: narrow down dependencies
