+++ pilot/pkg/networking/util/util.go
@@ -47,6 +47,7 @@ import (
 	"istio.io/istio/pkg/config"
 	kubelabels "istio.io/istio/pkg/kube/labels"
 	"istio.io/istio/pkg/log"
+	pm "istio.io/istio/pkg/model"
 	"istio.io/istio/pkg/proto/merge"
 	"istio.io/istio/pkg/util/strcase"
 	"istio.io/istio/pkg/wellknown"
@@ -92,12 +93,12 @@ const (
 )
 
 // ALPNH2Only advertises that Proxy is going to use HTTP/2 when talking to the cluster.
-var ALPNH2Only = []string{"h2"}
+var ALPNH2Only = pm.ALPNH2Only
 
 // ALPNInMeshH2 advertises that Proxy is going to use HTTP/2 when talking to the in-mesh cluster.
 // The custom "istio" value indicates in-mesh traffic and it's going to be used for routing decisions.
 // Once Envoy supports client-side ALPN negotiation, this should be {"istio", "h2", "http/1.1"}.
-var ALPNInMeshH2 = []string{"istio", "h2"}
+var ALPNInMeshH2 = pm.ALPNInMeshH2
 
 // ALPNInMeshH2WithMxc advertises that Proxy is going to use HTTP/2 when talking to the in-mesh cluster.
 // The custom "istio" value indicates in-mesh traffic and it's going to be used for routing decisions.
@@ -235,16 +236,7 @@ func SortVirtualHosts(hosts []*route.VirtualHost) {
 
 // ConvertLocality converts '/' separated locality string to Locality struct.
 func ConvertLocality(locality string) *core.Locality {
-	if locality == "" {
-		return &core.Locality{}
-	}
-
-	region, zone, subzone := label.SplitLocalityLabel(locality)
-	return &core.Locality{
-		Region:  region,
-		Zone:    zone,
-		SubZone: subzone,
-	}
+	return pm.ConvertLocality(locality)
 }
 
 // LocalityToString converts Locality struct to '/' separated locality string.
@@ -521,16 +513,7 @@ func IsAllowAnyOutbound(node *model.Proxy) bool {
 }
 
 func StringToExactMatch(in []string) []*matcher.StringMatcher {
-	if len(in) == 0 {
-		return nil
-	}
-	res := make([]*matcher.StringMatcher, 0, len(in))
-	for _, s := range in {
-		res = append(res, &matcher.StringMatcher{
-			MatchPattern: &matcher.StringMatcher_Exact{Exact: s},
-		})
-	}
-	return res
+	return pm.StringToExactMatch(in)
 }
 
 func StringToPrefixMatch(in []string) []*matcher.StringMatcher {
