+++ pkg/bootstrap/instance.go
@@ -25,9 +25,9 @@ import (
 
 	"github.com/Masterminds/sprig/v3"
 
-	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pkg/env"
 	"istio.io/istio/pkg/log"
+	"istio.io/istio/pkg/model"
 )
 
 const (
