+++ pkg/bootstrap/option/instances_test.go
@@ -23,8 +23,8 @@ import (
 
 	meshAPI "istio.io/api/mesh/v1alpha1"
 	networkingAPI "istio.io/api/networking/v1alpha3"
-	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pkg/bootstrap/option"
+	"istio.io/istio/pkg/model"
 )
 
 // nolint: lll
